+++ CHANGELOG.md
@@ -12,6 +12,7 @@ incremented upon a breaking change and the patch version will be incremented for
 
 **Added**
 
+- add `AccountDiscriminator` trait to derive discriminator for account types ([443](https://github.com/Ackee-Blockchain/trident/pull/443))
 - add more stake program related methods ([441](https://github.com/Ackee-Blockchain/trident/pull/441))
 - add remove functionality to AddressStorage ([442](https://github.com/Ackee-Blockchain/trident/pull/442))